+++ MakeYouPro.Bourse.Rates.Dal/Models/CNYDto.cs
@@ -15,13 +15,13 @@ namespace MakeYouPro.Bourse.Rates.Dal.Models
         [Required]
         public int Id { get; set; }
 
-        public double CNYRUB { get; set; }
-        public double CNYUSD { get; set; }
-        public double CNYEUR { get; set; }
-        public double CNYJPY { get; set; }
-        public double CNYRSD { get; set; }
-        public double CNYBGN { get; set; }
-        public double CNYARS { get; set; }
+        public decimal CNYRUB { get; set; }
+        public decimal CNYUSD { get; set; }
+        public decimal  CNYEUR { get; set; }
+        public decimal CNYJPY { get; set; }
+        public decimal CNYRSD { get; set; }
+        public decimal CNYBGN { get; set; }
+        public decimal CNYARS { get; set; }
 
         [Required]
         public DateTime dateTime { get; set; }
