+++ MakeYouPro.Bourse.Rates.Dal/Models/RSDDto.cs
@@ -15,13 +15,13 @@ namespace MakeYouPro.Bourse.Rates.Dal.Models
         [Required]
         public int Id { get; set; }
 
-        public double RSDRUB { get; set; }
-        public double RSDUSD { get; set; }
-        public double RSDEUR { get; set; }
-        public double RSDCNY { get; set; }
-        public double RSDJPY { get; set; }
-        public double RSDBGN { get; set; }
-        public double RSDARS { get; set; }
+        public decimal RSDRUB { get; set; }
+        public decimal RSDUSD { get; set; }
+        public decimal RSDEUR { get; set; }
+        public decimal RSDCNY { get; set; }
+        public decimal RSDJPY { get; set; }
+        public decimal RSDBGN { get; set; }
+        public decimal RSDARS { get; set; }
 
 
         [Required]
